+++ src/main/java/racingcar/input/utils/carname/validators/TwoOrMoreCarNamesValidatorUtils.java
@@ -1,4 +1,4 @@
-package racingcar.input.utils.carname;
+package racingcar.input.utils.carname.validators;
 
 import java.util.List;
 import racingcar.input.utils.carname.exception.LessThanTwoCarNamesException;
