+++ src/test/java/racingcar/input/utils/carname/CarNameFormatValidatorUtilsTest.java
@@ -7,6 +7,7 @@ import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import racingcar.input.utils.carname.exception.InvalidCarNameFormatException;
+import racingcar.input.utils.carname.validators.CarNameFormatValidatorUtils;
 
 public class CarNameFormatValidatorUtilsTest {
